Opacity button changes opacity of pen, borders button toggles grid borders
Two new buttons.
Clicking the opacity button uses black at a low opacity and uses that to fill squares. When square already has a background color that has an alpha less than 1, the opacity goes up by 10% every time the mouse passes over it again. This does not currently work with the rainbow button.
The borders button toggles the borders of the grid on and off.